fix: update metadata
